pyproject.toml
[project]
name = "dev-py-signup-gateway"
version = "0.0.1"
description = ""
authors = [
    {name = "ricardomarzochi07",email = "ricardomarzochi@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "buddybet-logmon-common (>=0.1.0,<0.2.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "python-jose (>=3.5.0,<4.0.0)",
    "cryptography (>=45.0.7,<46.0.0)",
    "buddybet-i18n (==0.6.0)",
    "buddybet-transactionmanager (==0.6.1)",
    "redis (>=7.0.1,<8.0.0)",
    "buddybet-idpsecure (==1.0.8)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
buddybet-i18n = {source = "pypi"}
buddybet-transactionmanager = {source = "pypi"}
